+++ extras/src/main/java/org/realityforge/arez/extras/Watcher.java
@@ -3,7 +3,7 @@ package org.realityforge.arez.extras;
 import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.jetbrains.annotations.TestOnly;
+import org.realityforge.anodoc.TestOnly;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.ComputedValue;
 import org.realityforge.arez.Node;